rename(route-folder): 채팅방 const 컨벤션 적용
