+++ src/app/(route)/chat/[postId]/_components/ChatRoomHeaderInfoButton/ChatRoomHeaderInfoButton.tsx
@@ -3,7 +3,7 @@
 import { Icon, ReportModal } from "@/components";
 import { cn } from "@/utils";
 import { useState } from "react";
-import { INFO_OPTIONS } from "./_internal/INFO_OPTIONS";
+import { INFO_OPTIONS } from "../CHATROOM_CONST";
 import { InfoButtonOptionValue } from "./_internal/InfoButtonOptionValueTypes";
 import useLeaveChatRoom from "@/api/fetch/chatRoom/api/useLeaveChatRoom";
 import { useClickOutside } from "@/hooks";
