+++ src/app/(route)/chat/[postId]/_components/ChatRoomMain/_internal/ChatBox/ChatBox.tsx
@@ -1,5 +1,5 @@
 import { cn, formatChatTime } from "@/utils";
-import { CHAT_SENDER_STYLE } from "../../constants/CHAT_SENDER_STYLE";
+import { CHAT_SENDER_STYLE } from "../../../CHATROOM_CONST";
 import ChatImageBox from "../ChatImageBox/ChatImageBox";
 import { ChatMessage } from "@/api/fetch/chatMessage/types/ChatMessageResponse";
 import { useGetUsersMe } from "@/api/fetch/user";
