--- src/app/(route)/chat/[postId]/_components/ChatRoomMain/constants/CHAT_SENDER_STYLE.ts
@@ -1,15 +1,0 @@
-export const CHAT_SENDER_STYLE = {
-  me: {
-    container: "justify-end",
-    timeOrder: "order-1",
-    bubbleOrder: "order-2",
-    bubbleColor:
-      "bg-fill-brand-normal-default bg-opacity-70 text-brand-subtle-default rounded-br-[0px]",
-  },
-  other: {
-    container: "justify-start",
-    timeOrder: "order-2",
-    bubbleOrder: "order-1",
-    bubbleColor: "bg-white text-layout-header-default rounded-bl-[0px]",
-  },
-};
